@@ -1,6 +1,6 @@
 {
-  "name": "drp",
-  "version": "1.0.0",
+  "name": "drp-back",
+  "version": "0.0.1",
   "description": "",
   "main": "index.js",
   "scripts": {
@@ -10,34 +10,27 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "aws-serverless-express": "^3.1.3",
     "aws-xray-sdk": "^1.2.0",
     "base-64": "^0.1.0",
     "batchelor": "^2.0.2",
-    "body-parser": "^1.18.2",
     "cheerio": "^1.0.0-rc.2",
-    "connect-dynamodb-session": "^1.1.0",
     "cookie": "^0.3.1",
-    "dateformat": "^2.0.0",
+    "dateformat": "^3.0.3",
     "escape-quotes": "^1.0.2",
-    "express": "^4.16.3",
-    "express-session": "^1.15.6",
-    "google-auth-library": "^0.10.0",
-    "googleapis": "^23.0.0",
-    "passport": "^0.4.0",
-    "passport-google-oauth": "^1.0.0",
-    "utf8": "^2.1.2"
+    "googleapis": "^28.1.0",
+    "utf8": "^3.0.0"
   },
   "devDependencies": {
-    "@types/aws-lambda": "0.0.12",
+    "@types/aws-lambda": "8.10.1",
     "@types/aws-sdk": "^2.7.0",
-    "@types/node": "^8.0.25",
+    "@types/node": "^9.6.4",
     "aws-sdk": "^2.104.0",
     "json-loader": "^0.5.7",
     "serverless-plugin-tracing": "^2.0.0",
-    "serverless-webpack": "^2.2.3",
-    "ts-loader": "^2.3.4",
+    "serverless-webpack": "^5.1.1",
+    "ts-loader": "^4.2.0",
     "typescript": "^2.4.2",
-    "webpack": "^3.6.0"
+    "webpack": "^4.6.0",
+    "webpack-node-externals": "^1.7.2"
   }
 }
