@@ -6,9 +6,9 @@ import {GetQueueAttributesRequest, QueueAttributeName,
 import {UpdateItemOutput} from 'aws-sdk/clients/dynamodb';
 
 
-import qu = require('./sub/_queue');
-import awsXRay = require('aws-xray-sdk');
-import awsPlain = require('aws-sdk');
+import * as qu from '../sub/_queue';
+import * as awsXRay from 'aws-xray-sdk';
+import * as awsPlain from 'aws-sdk';
 const AWS = awsXRay.captureAWS(awsPlain);
 const sqs: AWS.SQS = new AWS.SQS(),
       dynamo: AWS.DynamoDB.DocumentClient =  new AWS.DynamoDB.DocumentClient()
@@ -22,7 +22,7 @@ const TICKET_GENERATE_UNIT = Number(process.env.TICKET_GENERATE_UNIT)
  * チケット生成
  * @next -
  */
-export async function generateTicket(event, context, callback): Promise<void> {
+export const generateTicket = async (event, context, callback): Promise<void> => {
   console.log(JSON.stringify(event));
 
   let res: UpdateItemOutput;
@@ -45,7 +45,10 @@ export async function generateTicket(event, context, callback): Promise<void> {
     return;
   }
   console.log('ticket remain updated:' + JSON.stringify(res));
-
+  callback(null, {
+    "statusCode": 200,
+    "body": {}
+  });
 };
 
 
@@ -53,7 +56,7 @@ export async function generateTicket(event, context, callback): Promise<void> {
  * チケット消費
  * @next -
  */
-export async function consumeTicket(event: SNSEvent, context, callback): Promise<void> {
+export const consumeTicket = async (event: SNSEvent, context, callback): Promise<void> => {
   console.log(JSON.stringify(event));
 
   for(let rec of event.Records) {
@@ -81,4 +84,9 @@ export async function consumeTicket(event: SNSEvent, context, callback): Promise
     console.log('ticket remain updated:' + JSON.stringify(res));
 
   }
+
+  callback(null, {
+    "statusCode": 200,
+    "body": {}
+  });
 };
