@@ -1,9 +1,9 @@
 import {GetItemOutput, QueryOutput} from 'aws-sdk/clients/dynamodb';
 import {APIGatewayEvent} from 'aws-lambda';
-import {Job, JobStatus, CreateJobMessage, Session} from '../../types';
+import {Job, JobStatus, CreateJobMessage, Session} from './types';
 
-import awsXRay = require('aws-xray-sdk');
-import awsPlain = require('aws-sdk');
+import * as awsXRay from 'aws-xray-sdk';
+import * as awsPlain from 'aws-sdk';
 const AWS = awsXRay.captureAWS(awsPlain);
 const dynamo: AWS.DynamoDB.DocumentClient =  new AWS.DynamoDB.DocumentClient();
 
