@@ -1,24 +1,30 @@
 const path = require('path');
 const slsw = require('serverless-webpack');
+const nodeExternals = require('webpack-node-externals');
 
 module.exports = {
   entry: slsw.lib.entries,
   target: 'node',
+  mode: 'development',
+  optimization: {
+    minimize: false
+  },
   module: {
-    loaders: [
-      {test: /\.ts(x?)$/, loader: 'ts-loader' },
-      {test: /\.json$/, loader: 'json-loader' }
+    rules: [
+    // loaders: [
+      {
+        test: /\.ts$/,
+        loader: 'ts-loader',
+      }
     ]
   },
   resolve: {
-    extensions: ['.js', '.ts', '.tsx', '.jsx']
+    extensions: ['.ts', '.js', '.json']
   },
   output: {
     libraryTarget: 'commonjs',
     path: path.join(__dirname, '.built'),
     filename: 'handler.js'
   },
-  externals: {
-    "aws-sdk": 'commonjs aws-sdk',
-  }
+  externals: [ nodeExternals() ]
 };
